class end

+++ Platformer Game/Assets/Scripts/PlatformGenerator.cs
@@ -24,7 +24,7 @@ public class PlatformGenerator : MonoBehaviour
         while (true)
         {
             Vector3 newPosition = new Vector3(Random.Range(-6f, 6f), Random.Range(-3f, 3f), 0);
-            if(Mathf.Pow((transform.position.y - newPosition.y),2) < 9 && Mathf.Pow((transform.position.x - newPosition.x), 2) > 4)
+            if(Mathf.Pow((float)(transform.position.y - newPosition.y), 2) < 9 && Mathf.Pow((float)(transform.position.x - newPosition.x), 2) > 4)
             {
                 Debug.Log(transform.position);
                 Debug.Log("new platmor at " + newPosition);
